`scala` package is no longer needed
