+++ memeid4s/src/main/scala/memeid4s/digest/Digestible.scala
@@ -19,8 +19,8 @@ package memeid4s.digest
 import java.nio.charset.StandardCharsets.UTF_8
 
 import memeid.Bits.toBytes
-import memeid4s.scala.UUID
-import memeid4s.scala.UUID.RichUUID
+import memeid4s.UUID
+import memeid4s.UUID.RichUUID
 
 trait Digestible[A] { self =>
