+++ memeid4s/src/test/scala/memeid4s/SQUUIDSpec.scala
@@ -14,36 +14,26 @@
  * limitations under the License.
  */
 
-package memeid4s.scala
+package memeid4s
 
 import org.specs2.ScalaCheck
 import org.specs2.mutable.Specification
 
 @SuppressWarnings(Array("scalafix:Disable.scala.parallel"))
-class V4Spec extends Specification with ScalaCheck {
+class SQUUIDSpec extends Specification with ScalaCheck {
 
-  "V4 constructor" should {
+  "SQUUID constructor" should {
 
     "create version 4 UUIDs" in {
-      val uuids = (1 to 10).par.map(_ => UUID.V4.random.version)
+      val uuids = (1 to 10).par.map(_ => UUID.V4.squuid.version)
 
       uuids.to[Set] must contain(exactly(4))
     }
 
     "not generate the same UUID twice" in {
-      val uuids = (1 to 10).par.map(_ => UUID.V4.random)
+      val uuids = (1 to 10).par.map(_ => UUID.V4.squuid)
 
       uuids.to[Set].size must be equalTo 10
-    }
-
-    "be unable to create non-v4 values regardless of msb/lsb values provided" in {
-      val nonNull: UUID = UUID.V4(0, 0)
-      nonNull must not be equalTo(UUID.Nil)
-    }
-
-    "generate version 4 UUIDs regardless of msb/lsb values provided" in {
-      val nonNull: UUID = UUID.V4(0, 0)
-      nonNull.version must be equalTo 4
     }
 
   }
