+++ memeid4s/src/test/scala/memeid4s/V1Spec.scala
@@ -14,9 +14,9 @@
  * limitations under the License.
  */
 
-package memeid4s.scala
+package memeid4s
 
-import memeid4s.scala.UUID.RichUUID
+import memeid4s.UUID.RichUUID
 import org.specs2.ScalaCheck
 import org.specs2.mutable.Specification
