+++ memeid4s/src/test/scala/memeid4s/V3Spec.scala
@@ -14,14 +14,14 @@
  * limitations under the License.
  */
 
-package memeid4s.scala
+package memeid4s
 
 import org.specs2.ScalaCheck
 import org.specs2.mutable.Specification
 
-class V5Spec extends Specification with ScalaCheck {
+class V3Spec extends Specification with ScalaCheck {
 
-  "V5 constructor" should {
+  "V3 constructor" should {
 
     "create same UUID for the same namespace/name" in {
       val namespace = UUID.V4.random
